pyproject.toml
[project]
name = "apple-eventkit-mcp"
version = "0.1.0"
description = "MCP server for Apple Calendar and Reminders via EventKit"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "pyobjc-framework-EventKit>=10.0",
    "pyobjc-framework-Cocoa>=10.0",
]

[project.scripts]
apple-eventkit-mcp = "apple_eventkit_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/apple_eventkit_mcp"]

[tool.hatch.build.targets.sdist]
include = ["src/apple_eventkit_mcp"]

[tool.hatch.envs.default]
path = ".venv"

[tool.uv]
package = true

[tool.uv.sources]
